:recycle: game 서비스 로직 분리
