+++ src/main/java/com/danaleeband/guessit/game/GameSocketController.java
@@ -1,0 +1,18 @@
+package com.danaleeband.guessit.game;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.messaging.handler.annotation.DestinationVariable;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.stereotype.Controller;
+
+@Controller
+@RequiredArgsConstructor
+public class GameSocketController {
+
+    private final GameService gameService;
+
+    @MessageMapping("/rooms/{roomId}/start")
+    public void startGame(@DestinationVariable long roomId) {
+        gameService.startGame(roomId);
+    }
+}
