+++ src/main/java/com/danaleeband/guessit/room/controller/RoomSocketController.java
@@ -6,7 +6,6 @@ import com.danaleeband.guessit.room.dto.RoomLeaveRequestDto;
 import com.danaleeband.guessit.room.dto.RoomLeaveResponseDto;
 import com.danaleeband.guessit.room.dto.RoomSessionBindRequestDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
@@ -40,10 +39,5 @@ public class RoomSocketController {
     @MessageMapping("/rooms/ready")
     public void ready(@Payload GameReadyRequestDto gameReadyRequestDto) {
         roomService.updatePlayerReady(gameReadyRequestDto);
-    }
-
-    @MessageMapping("/rooms/{roomId}/start")
-    public void startGame(@DestinationVariable long roomId) throws InterruptedException {
-        roomService.startGame(roomId);
     }
 }
